move to `chia/types/hints.py`
