+++ chia/server/start_introducer.py
@@ -4,11 +4,11 @@ import pathlib
 import sys
 from typing import Any, Dict, Optional
 
-from chia.hints import IntroducerService
 from chia.introducer.introducer import Introducer
 from chia.introducer.introducer_api import IntroducerAPI
 from chia.server.outbound_message import NodeType
 from chia.server.start_service import Service, async_run
+from chia.types.hints import IntroducerService
 from chia.util.chia_logging import initialize_service_logging
 from chia.util.config import load_config, load_config_cli
 from chia.util.default_root import DEFAULT_ROOT_PATH
