+++ chia/simulator/setup_services.py
@@ -14,16 +14,6 @@ from typing import Any, AsyncGenerator, AsyncIterator, Dict, Iterator, List, Opt
 from chia.cmds.init_funcs import init
 from chia.consensus.constants import ConsensusConstants
 from chia.daemon.server import WebSocketServer, daemon_launch_lock_path
-from chia.hints import (
-    CrawlerService,
-    FarmerService,
-    FullNodeService,
-    HarvesterService,
-    IntroducerService,
-    SimulatorFullNodeService,
-    TimelordService,
-    WalletService,
-)
 from chia.protocols.shared_protocol import Capability, capabilities
 from chia.seeder.dns_server import DNSServer, create_dns_server_service
 from chia.seeder.start_crawler import create_full_node_crawler_service
@@ -41,6 +31,16 @@ from chia.simulator.start_simulator import create_full_node_simulator_service
 from chia.ssl.create_ssl import create_all_ssl
 from chia.timelord.timelord_launcher import VDFClientProcessMgr, find_vdf_client, spawn_process
 from chia.types.blockchain_format.sized_bytes import bytes32
+from chia.types.hints import (
+    CrawlerService,
+    FarmerService,
+    FullNodeService,
+    HarvesterService,
+    IntroducerService,
+    SimulatorFullNodeService,
+    TimelordService,
+    WalletService,
+)
 from chia.types.peer_info import UnresolvedPeerInfo
 from chia.util.bech32m import encode_puzzle_hash
 from chia.util.config import config_path_for_filename, load_config, lock_and_load_config, save_config, set_peer_info
