+++ chia/simulator/start_simulator.py
@@ -8,13 +8,13 @@ from pathlib import Path
 from typing import Any, Dict, List, Optional, Tuple
 
 from chia.full_node.full_node import FullNode
-from chia.hints import SimulatorFullNodeService
 from chia.server.outbound_message import NodeType
 from chia.server.start_service import Service, async_run
 from chia.simulator.block_tools import BlockTools, test_constants
 from chia.simulator.full_node_simulator import FullNodeSimulator
 from chia.simulator.simulator_full_node_rpc_api import SimulatorFullNodeRpcApi
 from chia.types.blockchain_format.sized_bytes import bytes32
+from chia.types.hints import SimulatorFullNodeService
 from chia.util.bech32m import decode_puzzle_hash
 from chia.util.chia_logging import initialize_logging
 from chia.util.config import load_config, load_config_cli, override_config
