+++ tests/farmer_harvester/test_farmer_harvester.py
@@ -10,7 +10,6 @@ from chia_rs import G1Element
 
 from chia.cmds.cmds_util import get_any_service_client
 from chia.farmer.farmer import Farmer
-from chia.hints import FarmerService, HarvesterService
 from chia.plotting.util import PlotsRefreshParameter
 from chia.protocols import farmer_protocol, harvester_protocol
 from chia.protocols.protocol_message_types import ProtocolMessageTypes
@@ -18,6 +17,7 @@ from chia.rpc.harvester_rpc_client import HarvesterRpcClient
 from chia.server.outbound_message import NodeType, make_msg
 from chia.simulator.block_tools import BlockTools
 from chia.types.blockchain_format.sized_bytes import bytes32
+from chia.types.hints import FarmerService, HarvesterService
 from chia.types.peer_info import UnresolvedPeerInfo
 from chia.util.config import load_config
 from chia.util.hash import std_hash
