+++ tests/plot_sync/test_plot_sync.py
@@ -13,7 +13,6 @@ from chia_rs import G1Element
 
 from chia.farmer.farmer import Farmer
 from chia.harvester.harvester import Harvester
-from chia.hints import FarmerService, HarvesterService
 from chia.plot_sync.delta import Delta, PathListDelta, PlotListDelta
 from chia.plot_sync.receiver import Receiver
 from chia.plot_sync.sender import Sender
@@ -24,6 +23,7 @@ from chia.protocols.harvester_protocol import Plot
 from chia.protocols.protocol_message_types import ProtocolMessageTypes
 from chia.simulator.block_tools import BlockTools
 from chia.types.blockchain_format.sized_bytes import bytes32
+from chia.types.hints import FarmerService, HarvesterService
 from chia.util.config import create_default_chia_config, lock_and_load_config, save_config
 from chia.util.ints import uint8, uint32, uint64
 from chia.util.misc import SplitAsyncManager, split_async_manager
